improvement(dashboard-layout): move type declarations to the correspondent component 1-) Moved ProfileData type to SideDrawerProfile 2-) Moved MenuItem type to SideDrawerMenu.tsx
