--- web-client/src/types/profile-data.d.ts
@@ -1,6 +1,0 @@
-export interface ProfileData {
-  displayName: string;
-  displayPicture: string;
-  casesCompleted: number;
-  averageRating: number;
-}
